+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -91,6 +91,7 @@
 #import "SentryCrashReportSink.h"
 #import "SentryCrashReportStore.h"
 #import "SentryCrashScopeObserver.h"
+#import "SentryCrashStackCursor_SelfThread.h"
 #import "SentryCrashStackEntryMapper.h"
 #import "SentryCrashUUIDConversion.h"
 #import "SentryCrashWrapper.h"