+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Crash in setMeasurement when name is nil (#5064)
 - Make setMeasurement thread safe (#5067, #5078)
 
+### Improvements
+
+- Refactor session replay types and logging
+
 ## 8.49.0
 
 ### Features