+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Crash in setMeasurement when name is nil (#5064)
 - Make setMeasurement thread safe (#5067, #5078)
+- Async safe log for backtrace in CPPException (#5098)
 
 ## 8.49.0