+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Make setMeasurement thread safe (#5067, #5078)
 - Truncation of Swift crash messages (#5036)
 - Add error logging for move current replay to last path (#5083)
+- Async safe log for backtrace in CPPException (#5098)
 
 ## 8.49.0